--- app/src/main/res/layout/activity_main.xml
@@ -1,24 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".MainActivity">
-
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:elevation="8dp"
-        style="@style/ToolbarTheme" />
-
-    <ListView
-        android:id="@+id/main_menu"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:divider="@color/transparent"
-        android:dividerHeight="0dp"
-        android:listSelector="@android:color/transparent"
-        android:background="@color/off_white" />
-
-</LinearLayout>
